Untangle ratelimiter quota from python flag
