+++ crates/adapters/coinbase_intx/src/http/client.rs
@@ -22,6 +22,7 @@
 
 use std::{
     collections::HashMap,
+    num::NonZeroU32,
     sync::{Arc, LazyLock, Mutex},
 };
 
@@ -85,7 +86,7 @@ pub struct CoinbaseIntxResponse<T> {
 
 // https://docs.cdp.coinbase.com/intx/docs/rate-limits#rest-api-rate-limits
 pub static COINBASE_INTX_REST_QUOTA: LazyLock<Quota> =
-    LazyLock::new(|| Quota::rate_per_second(40).unwrap());
+    LazyLock::new(|| Quota::per_second(NonZeroU32::new(40).unwrap()));
 
 /// Provides a lower-level HTTP client for connecting to the [Coinbase International](https://coinbase.com) REST API.
 ///
