+++ crates/network/src/ratelimiter/quota.rs
@@ -16,7 +16,6 @@
 use std::{num::NonZeroU32, prelude::v1::*, time::Duration};
 
 use nonzero_ext::nonzero;
-use pyo3::{exceptions::PyException, prelude::*};
 
 use super::nanos::Nanos;
 
@@ -50,39 +49,6 @@ use super::nanos::Nanos;
 pub struct Quota {
     pub(crate) max_burst: NonZeroU32,
     pub(crate) replenish_1_per: Duration,
-}
-
-#[pymethods]
-impl Quota {
-    #[staticmethod]
-    pub fn rate_per_second(max_burst: u32) -> PyResult<Self> {
-        match NonZeroU32::new(max_burst) {
-            Some(max_burst) => Ok(Self::per_second(max_burst)),
-            None => Err(PyErr::new::<PyException, _>(
-                "Max burst capacity should be a non-zero integer",
-            )),
-        }
-    }
-
-    #[staticmethod]
-    pub fn rate_per_minute(max_burst: u32) -> PyResult<Self> {
-        match NonZeroU32::new(max_burst) {
-            Some(max_burst) => Ok(Self::per_minute(max_burst)),
-            None => Err(PyErr::new::<PyException, _>(
-                "Max burst capacity should be a non-zero integer",
-            )),
-        }
-    }
-
-    #[staticmethod]
-    pub fn rate_per_hour(max_burst: u32) -> PyResult<Self> {
-        match NonZeroU32::new(max_burst) {
-            Some(max_burst) => Ok(Self::per_hour(max_burst)),
-            None => Err(PyErr::new::<PyException, _>(
-                "Max burst capacity should be a non-zero integer",
-            )),
-        }
-    }
 }
 
 /// Constructors for Quotas
